+++ api/schedules/status_blockchains.py
@@ -20,9 +20,6 @@ from api import app
 from api import utils
 
 def update():
-    if not globals.farming_enabled():
-        #app.logger.info("Skipping blockchains status collection on non-farming instance.")
-        return
     with app.app_context():
         try:
             blockchains = ['chia']