+++ api/schedules/status_controller.py
@@ -23,8 +23,6 @@ from api import app
 from api import utils
 
 def update():
-    if not utils.is_controller():
-        return
     with app.app_context():
         try:
             workers = db.session.query(w.Worker).order_by(w.Worker.hostname).all()