+++ api/views/plots/resources.py
@@ -26,28 +26,15 @@ class Plots(MethodView):
     @blp.response(200, PlotSchema(many=True))
     @blp.paginate(SQLCursorPage)
     def get(self, args):
-        ret = Plot.query.filter_by(**args)
+        ret = db.session.query(Plot).filter_by(**args)
         return ret
 
     @blp.etag
     @blp.arguments(BatchOfPlotSchema)
     @blp.response(201, PlotSchema(many=True))
     def post(self, new_items):
-        # Get plot info via RPC to determine type: solo or portable
-        plots_via_rpc = chia.get_all_plots()
-        # Now delete all old plots by hostname of first new plotting
-        db.session.query(Plot).filter(Plot.hostname==new_items[0]['hostname']).delete()
+        # Reworked status_plots.py to load directly on Controller after Flax updated to v0.1.1
         items = []
-        for new_item in new_items:
-            item = Plot(**new_item)
-            item.type = ""
-            for plot_rpc in plots_via_rpc:
-                if plot_rpc['plot_id'].startswith("0x{0}".format(item.plot_id)) and 'type' in plot_rpc:
-                    item.type = plot_rpc['type']
-                    #app.logger.info("Found type: {0}".format(item.type))
-            db.session.add(item)
-            items.append(item)
-        db.session.commit()
         return items
 
 
@@ -63,14 +50,8 @@ class PlotByHostname(MethodView):
     @blp.arguments(BatchOfPlotSchema)
     @blp.response(200, PlotSchema(many=True))
     def put(self, new_items, hostname):
-        # Now delete all old plots by hostname of first new plotting
-        db.session.query(Plot).filter(Plot.hostname==hostname).delete()
+        # Reworked status_plots.py to load directly on Controller after Flax updated to v0.1.1
         items = []
-        for new_item in new_items:
-            item = Plot(**new_item)
-            items.append(item)
-            db.session.add(item)
-        db.session.commit()
         return items
 
     @blp.etag